fix(scion-rcp-plugin): reject messages from untrusted origins
Since we are bridging messages from the client to the host, the "SCION Microfrontend Platform" host cannot perform the "actual" origin check as we dispatch messages to the host under the host's origin. For that reason, we need to check the "actual" origin prior to dispatching the message to the host.

Also, when registering the applications, we set the applications "actual" message origin to the origin of the host as SCION would reject messages otherwise.