feat(scion-rcp-plugin): externalize JavaScript into separate files
Since the library must be compatible with Java 11, we cannot use text blocks for multi-line strings. For this reason, we have externalized JavaScript into separate files.

We have also changed the placeholder syntax from "${PLACEHOLDER}" to "/@@PLACEHOLDER@@/" to be valid JavaScript and not to confuse with the JavaScript template literal placeholder syntax.