fix(scion-rcp-plugin): support conversion of JavaScript dictionaries into Java Maps
We use a custom GSON adapter to convert a Java `Map` to a JSON object literal and vice versa, which is necessary because JSON has no representation for the Map data type.
By default, GSON converts a Java `Map` to a JSON object literal, so there is no way to determine in JavaScript whether it is a Map or a dictionary.

Format of the JSON object literal: `{__type: 'Map', __value: [...[key, value]]}`